Reorganize import code to make is modular
